moved destination diagnostics
