+++ mainnet_launch/pages/autopool_crm/weighted_crm.py
@@ -21,18 +21,6 @@ def fetch_and_render_weighted_crm_data(autopool: AutopoolConstants):
         )
 
 
-def fetch_and_render_destination_apr_data(autopool: AutopoolConstants):
-    apr_components_fig = _make_apr_components_fig(autopool)
-    st.plotly_chart(apr_components_fig, use_container_width=True)
-
-    with st.expander("See explanation"):
-        st.write(
-            f"""
-              APR Components: Show Unweighted Base, Fee, Incentive and Price Return of the Destination
-            """
-        )
-
-
 def _make_all_destination_composite_return_df(autopool: AutopoolConstants) -> go.Figure:
     compositeReturn_out_df = 100 * fetch_destination_summary_stats(autopool, "compositeReturn")
     pricePerShare_df = 100 * fetch_destination_summary_stats(autopool, "pricePerShare")
@@ -50,33 +38,6 @@ def _make_all_destination_composite_return_df(autopool: AutopoolConstants) -> go
         line=dict(dash="dash", color="blue"),
     )
     return composite_return_fig
-
-
-def _make_apr_components_fig(autopool: AutopoolConstants) -> go.Figure:
-    priceReturn_df = 100 * fetch_destination_summary_stats(autopool, "priceReturn")
-    baseApr_df = 100 * fetch_destination_summary_stats(autopool, "baseApr")
-    feeApr_df = 100 * fetch_destination_summary_stats(autopool, "feeApr")
-    incentiveApr_df = 100 * fetch_destination_summary_stats(autopool, "incentiveApr")
-    pointsApr_df = 100 * fetch_destination_summary_stats(autopool, "pointsApr")
-
-    st.title("Destination APR Components")
-
-    destination = st.selectbox("Select a destination", pointsApr_df.columns)
-
-    plot_data = pd.DataFrame(
-        {
-            "Price Return": priceReturn_df[destination],
-            "Base APR": baseApr_df[destination],
-            "Incentive APR": incentiveApr_df[destination],
-            "Fee APR": feeApr_df[destination],
-            "Points APR": pointsApr_df[destination],
-        },
-        index=pointsApr_df.index,
-    )
-
-    apr_components_fig = px.line(plot_data, title=f"APR Components for {destination}")
-    _apply_default_style(apr_components_fig)
-    return apr_components_fig
 
 
 def _apply_default_style(fig: go.Figure) -> None:
